job_submit.lua: manage multiple qos same settings
Previously, the job_submit.lua script was not able to handle multiple
QOS which share the same maxcpus and maxwall settings. Only the latest
QOS in qos.conf file was considered, the previous one being overriden
in qos_list hash. With this commit, the script maintains a list of QOS
with the same maxcpus/maxwall and iterate over it in the big main
nested loop.