pwmgt: daemonize stop wrapper cmd
This way, the SSH client launched on slurmctld server can exit *before*
is closed by the remote host due to the poweroff/suspend. This way, the
SSH client exits with the 0 return code instead of 255 and no error is
wrongly reported in the slurmctld server logs when the remote command is
successfully launched.

Fix #9