job_submit: split optimization
The split implementation was using a broad pattern ('(.-)<separator>').
This pattern takes a long time (tenths of second) to match on large
strings (few KB). The new implementation just search for the <separator>
and find the matched string with the indexes returned.

This suboptimal function was impacting the time to parse qos.conf when
there is a large number of account. a large number of account is the
norm for cluster that still uses one account per user.

This patch brings the execution of sbatch from 2~3 seconds to less than
a tenth of seconds on one cluster.