job_submit.lua: remove getent to get username
The job submit Lua script executed `getent` command to get the name
(or login) of the user who submitted the job, using its uid. Actually,
Slurm now provides the username in the Lua interface of the job
description. This commit now retrieves the username from this interface,
thus making the getent command useless and remove it.