job_submit.lua: replace cat + remove sacctmgr
Replace cat with Lua standard library io open, this is much more
effective and simpler. The commit also makes the Lua script fail if QOS
configuration file does not exist, it does not run sacctmgr as a
fallback solution anymore as this is both very bad for performances and
can DOS slurmdbd eventually with large jobs submissions. It is actually
better to make the submission fail in this case.