admin-utils nodes-only to slurm-check-procs-nojob
Add a --nodes-only output parameter for the output, will be used
by monitoring scripts.

Other output are moved to standard errors.

This also add a parameter to use an alternative path for the helper
scripts, this is usefull during tests and debugging.

This also includes a lot of small pep8 fixes.