genscripts: rework cgroup cleaning in epilog
There were few bugs in the way the epilog tried to remove slurm
cgroups:

- Only the cpuset controller was considered, but there are other
  controllers (eg. freezer, memory, etc).
- The job cgroup was tried to be removed before the step cgroup were
  removed, which lead to EBUSY. This is fixed using find -delete, which
  deletes folders in-depth first.
- The user cgroup was not removed.

All these 3 bugs are fixed.

Please note the list of controllers is hard-coded to cpuset and freezer
because only these 2 controllers are used by Slurm on EDF clusters.

The commit also mentions in comments this stuff is a work-around for old
Linux kernel bug, this must not be considered as a normal thing to do
in epilog.