job_submit: Exclude non matching QoS' names from partition validation
QoSes are expected to be delimited by '_', the first field being the
name of a partition.
In the 'QoS provided' case, if:
- we have no partition for the job and the QoS didn't had '_', then we
  cannot find the appropriate partition: we bail out
- we have a partition for the job and we could deduce the partition from
  the QoS: we will make sure they match.

Said differently, for QoSes which don't follow the naming standard, the
partition is expected to be given when submitting a job, and it is not
checked against the QoS' name.