job_submit: Always assign a hard coded time limit when none is specified
We have always been assigning a 1h time limit to jobs which were
submitted without a QoS and without a time limit.

If no time limit is specified, slurm sets it to infinite.

Since we introduced the time limit validation to make sure a job's limit
is less or equal than the QoS' limit, we have an issue where jobs
submitted with a QoS but no time limit have an infinite time limit.

As a result, they all fail the time limit validation. Based on users'
feedback, it has been decided to assign a default 1h time limit as well
in that case to make the whole thing clearer.