wckeys: fix initial insertion on empty SQL table
The INSERT SQL request did not work when the SlurmDBD MySQL wckey table
is empty because the subrequest in the NOT IN test did not return any
result and this produced an error eventually. The INSERT request has
been reworked using NOT EXISTS test that handles empty tables properly.

Unfortunately, more native solutions such as REPLACE and INSERT IGNORE
cannot be adopted as the wckey_name is not a primary key on the table.

The script is now able to insert wckeys in empty SlurmDBD database
cluster wckeys table.

Fix #8