wckeys: validate project/app codes length
The format of projects and applications codes found in downloaded files
are validated against a maximum string length. If the code is greather
than the defined maximum length, the script fails with an error. The
script provides defaults values for maximum lengths which can be
overriden in /etc/{default,sysconfig}/wckeyctl environment file.