epilog/clean: Compatibility with cgroupv2 on EL9
Refactor the script to be compatible with cgroupv2. This introduce an
RPC call, but could not find a way around that with the current cgroup
files hierarchy.

Also removed historic psm shm handling (dating from the qlogic days).

Removed special handling of the private tmp dir spank plugin, not really
used anymore and the generic case should work just fine.